requirements.txt
django>=4.2,<5
mysqlclient>=2.0.0
django-import-export>=3.3
